[Component] create module instance based on export list
The problem is from binary definition, I cannot see how it stores sort:idx. Thus, the current implementation about `exportXxx` is wrong since the sort is not in the immediate module instance but somewhere else. On the other hand, the immediate module instance is correct semantic.

To fix this problem, I will propose a test and see how the tool generates reference to sort(like a function instantiate in another module), and check what should be added into the internal interface to interact with these.

Signed-off-by: Lîm Tsú-thuàn <dannypsnl@secondstate.io>