[Utils] Install Script: Support Ubuntu packages for versions above o.11.1
GitHub Issue Reference: https://github.com/WasmEdge/WasmEdge/issues/2397

From WasmEdge 0.11.1, we have the Ubuntu release.
For the WasmEdge-tensorflow and WasmEdge-image, they have
updated the ubuntu releases.
This commit installs the same DIST of the extensions and WasmEdge,
such as the plugins (the policy described in the comment of #2196).

For the WasmEdge-tensorflow-tools, we do not apply this because
the tools are the stand-alone executables.

The extension with Ubuntu releases can be test: 0.11.1, 0.11.2,
and 0.12.0-alpha.2

Signed-off-by: Shreyas Atre <shreyasatre16@gmail.com>