[CMake] Fix version checking in source tarball
In source tarball, .git directory does not exist. So, git describe
command fails with the following error.

fatal: not a git repository (or any of the parent directories): .git

This commit checks if the current source directory is a git repository.
Otherwise, cmake does not run the git describe command.

Signed-off-by: Biswapriyo Nath <nathbappai@gmail.com>