[CMake] Fix the .git Detection Issue (#3635)
In #3626, .git detection was introduced to avoid the error
triggered by the git describe command. However, the
GIT_VERSION_NOT_FOUND variable is defined and set within the git
describe command call. After applying #3626, GIT_VERSION_NOT_FOUND
will never be defined and set if the git command exists but the
.git folder does not.

In this PR, we define and initialize GIT_VERSION_NOT_FOUND to 1.
This will ensure that CPACK_PACKAGE_VERSION is set to
"0.0.0-unreleased" if git describe is not working or the .git
folder does not exist.

Signed-off-by: hydai <z54981220@gmail.com>